Complete migration to player_components_factory.cc
